add highpass filter

+++ filter/lowpass_filter/lowpass_filter.h
@@ -7,19 +7,19 @@ class LowPassFilter : public Filter {
 
 public:
     
-    LowPassFilter(float cutFrq, uint32_t sampleFrq); // 低通滤波器截止频率
-    LowPassFilter(float cof); // 低通滤波器系数
+    LowPassFilter(float cutFrq, float sampleFrq);
+    LowPassFilter(float cof);
     ~LowPassFilter();
 
-    void run(int16_t *Vi, int16_t *Vo);
+    void run(float *Vi, float *Vo);
     
 
 private:
 
-    uint32_t _sampleFrq;
+    float _sampleFrq;
     float _cutFrq;
     float _cof;
-    int16_t _Vo_p;
+    float _Vo_p;
     bool _direct;
 
 };
